Progress, almost working but still needs to fix vulkan yap errors

+++ Engine/src/Resource/Mesh.cpp
@@ -29,7 +29,7 @@ void Mesh::CreateFrom(float *vertices, uint32_t verticeCount, uint32_t *indices,
         std::copy(indices, indices + indiceCount, m_indices.data());
     }
     SubMesh m;
-    m.count = m_vertices.size();
+    m.count = (uint32_t)m_vertices.size();
     m.startIndex = 0;
     m_subMeshes.clear();
     m_subMeshes.push_back(m);
